Minor fixes.

+++ mg_visual music/NoteStyle_Line.cs
@@ -515,7 +515,7 @@ namespace Visual_Music
 			if ((float)(nextNote.start - note.stop) > Qn_gapThreshold * Project.Notes.TicksPerBeat || note == nextNote)
 			{
 				if (nextNoteStart.X != noteStart.X)
-					nextNoteStart.Y = (int)MathHelper.Lerp(noteStart.Y, nextNoteStart.Y, (float)(noteEnd - noteStart.X) / (nextNoteStart.X - noteStart.X));
+					nextNoteStart.Y = MathHelper.Lerp(noteStart.Y, nextNoteStart.Y, (float)(noteEnd - noteStart.X) / (nextNoteStart.X - noteStart.X));
 				nextNoteStart.X = noteEnd;
 			}
